pyproject.toml
[project]
name = "crawler-hub"
version = "0.1.0"
description = "AI 短剧情报系统：爬全网最热 AI 短视频 + 最新 AI 工具，每周一豆包摘要 + 微信推送"
requires-python = ">=3.11"

dependencies = [
    "httpx>=0.27.0",
    "huggingface-hub>=0.24.0",
    "bilibili-api-python>=17.0.0",
    "google-api-python-client>=2.140.0",
    "yt-dlp>=2024.10.0",
    "openai>=1.40.0",        # 豆包接口走 OpenAI 兼容
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "tenacity>=9.0.0",       # 通用重试
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[tool.setuptools.packages.find]
where = ["."]
include = ["sources*", "analyzers*", "storage*", "reporter*", "notifier*"]
